Create UnityHand.cs
I found this a little hard to understand but it is the part of the boxer scene that moves the hands I believe.  There is one of these scripts on each of the gloves.  I'm confused about a lot of the stuff here so maybe we should go over it during a class period.